requirements.txt
gradio==4.44.0
langchain==0.2.11
langchain-community==0.2.10
chromadb==0.4.24
pypdf==4.3.1
pydantic==2.9.1
python-dotenv
openai
